test(お知らせ): バナー選定の純関数を Firebase 非依存の src/lib へ移し、テストが SDK 初期化を踏まないようにする
